renaming module path
